+++ CHANGELOG.md
@@ -11,6 +11,7 @@ project adheres to
 
 * Current stable rust is 1.57, MSRV is now 1.46.0.
 * Update optional rsass to 0.23.0.
+* Update env_logger to 0.9 and gotham to 0.7.1 in examples
 
 
 ## Release 0.13.4 - 2021-06-25